@@ -15,31 +15,39 @@
 
 package com.keylesspalace.tusky;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.graphics.Color;
 import android.media.MediaPlayer;
+import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
+import android.view.MotionEvent;
+import android.view.View;
 import android.widget.MediaController;
+import android.widget.ProgressBar;
 import android.widget.VideoView;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-
 public class ViewVideoActivity extends BaseActivity {
-    @BindView(R.id.video_player) VideoView videoView;
-    @BindView(R.id.toolbar) Toolbar toolbar;
+
+    Handler handler = new Handler(Looper.getMainLooper());
+    Toolbar toolbar;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_video);
-        ButterKnife.bind(this);
 
+        final ProgressBar progressBar = findViewById(R.id.video_progress);
+        VideoView videoView = findViewById(R.id.video_player);
+
+        toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-
         ActionBar bar = getSupportActionBar();
-
         if (bar != null) {
             bar.setTitle(null);
             bar.setDisplayHomeAsUpEnabled(true);
@@ -56,10 +64,30 @@ public class ViewVideoActivity extends BaseActivity {
         videoView.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
             @Override
             public void onPrepared(MediaPlayer mp) {
+                progressBar.setVisibility(View.GONE);
                 mp.setLooping(true);
+                hideToolbarAfterDelay();
             }
         });
         videoView.start();
+
+        videoView.setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                if (event.getAction() == MotionEvent.ACTION_DOWN) {
+                    handler.removeCallbacksAndMessages(null);
+                    toolbar.animate().cancel();
+                    toolbar.setAlpha(1);
+                    toolbar.setVisibility(View.VISIBLE);
+                    hideToolbarAfterDelay();
+                }
+                return false;
+            }
+        });
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            getWindow().setStatusBarColor(Color.BLACK);
+        }
     }
 
     @Override
@@ -71,5 +99,23 @@ public class ViewVideoActivity extends BaseActivity {
             }
         }
         return super.onOptionsItemSelected(item);
+    }
+
+    void hideToolbarAfterDelay() {
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                toolbar.animate().alpha(0).setListener(new AnimatorListenerAdapter() {
+                    @Override
+                    public void onAnimationEnd(Animator animation) {
+                        View decorView = getWindow().getDecorView();
+                        int uiOptions = View.SYSTEM_UI_FLAG_LOW_PROFILE;
+                        decorView.setSystemUiVisibility(uiOptions);
+                        toolbar.setVisibility(View.INVISIBLE);
+                        animation.removeListener(this);
+                    }
+                });
+            }
+        }, 3000);
     }
 }
